Replaced sme enum classes with records, entities and services | Implemented some business rules

+++ sources/core/Ggio.BikeSherpa.Backend.Domain/DeliveryAggregate/PackingSize.cs
@@ -5,6 +5,6 @@ public record PackingSize(
      int MaxWeight,
      int TourMaxLength,
      int MaxLength,
-     int TourPrice,
-     int Price
+     double TourPrice,
+     double Price
      );
